Use purescript-dom from contrib
purescript-dom only exports DOM types which are equivalent to those in
this library. Since its available from purescript-contrib, its pretty
widely used so this should use that version.

Without this change, mixing effectful code that uses
purescript-contrib's DOM and code from this library won't compile
because there will be a duplicate row in the effect monad.