Fix row name conflict in timeout functions
If you're running an effect that needs the DOM (which is pretty common),
the previous code would cause a compile error due to a duplicate effect
row `dom :: DOM`. This solves that case. The only way I could see this
being a problem is if you needed to fire callback code that *must not*
have access to the DOM, but I don't think that's a terribly reasonable
expectation since you need the DOM to even fire a callback.